+++ appveyor.ps1
@@ -5,12 +5,6 @@ echo "Downloading dub...";
 Invoke-WebRequest https://code.dlang.org/files/dub-1.0.0-beta.1-windows-x86.zip -OutFile c:\dub.zip;
 7z x c:\dub.zip -oc:\dub > $null;
 
-if ($env:arch -eq "x86") {
-  echo "Downloading 32-bit LLVM...";
-  Invoke-WebRequest http://llvm.org/releases/3.8.1/LLVM-3.8.1-win32.exe -OutFile c:\LLVM.exe;
-  7z x c:\LLVM.exe -oc:\PROGRA~2\LLVM > $null;
-}
-
 $env:PATH = "c:\dub;$($env:PATH)";
 $env:PATH = "c:\dmd2\windows\bin;$($env:PATH)";
 
@@ -18,7 +12,6 @@ $env:compilersetup = "C:\Program Files (x86)\Microsoft Visual Studio 14.0\VC\vcv
 
 if ($env:arch -eq "x86") {
   $env:compilersetupargs = "x86";
-  $env:PATH = "c:\PROGRA~2\LLVM\bin;$($env:PATH)";
   $env:archswitch = "";
 }
 else {
@@ -47,13 +40,6 @@ function Invoke-CmdScript {
 }
 
 Invoke-CmdScript $env:compilersetup $env:compilersetupargs;
-
-echo "Building dstep...";
-dub build $env:archswitch;
-
-if ($LastExitCode -ne 0) {
-  $host.SetShouldExit(-1)
-}
 
 echo "Running tests...";
 dub --config=test $env:archswitch;